requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
greenlet==3.3.0
idna==3.11
iniconfig==2.3.0
packaging==25.0
pillow==12.1.0
playwright==1.57.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dotenv==1.2.1
python-slugify==8.0.4
requests==2.32.5
schedule==1.2.2
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
